+++ fn/objective.go
@@ -2,6 +2,7 @@ package fn
 
 import (
 	"github.com/roidaradal/fn/ds"
+	"github.com/roidaradal/fn/list"
 	"github.com/roidaradal/opt/discrete"
 )
 
@@ -9,4 +10,9 @@ import (
 func ScoreCountUniqueValues(solution *discrete.Solution) discrete.Score {
 	uniqueValues := ds.SetFrom(solution.Values())
 	return discrete.Score(uniqueValues.Len())
+}
+
+// ScoreSubsetSize counts the size of solution as subset
+func ScoreSubsetSize(solution *discrete.Solution) discrete.Score {
+	return discrete.Score(list.Sum(solution.Values()))
 }