+++ fn/string.go
@@ -1,8 +1,12 @@
 package fn
 
 import (
+	"cmp"
+	"slices"
 	"strings"
 
+	"github.com/roidaradal/fn/list"
+	"github.com/roidaradal/fn/str"
 	"github.com/roidaradal/opt/discrete"
 )
 
@@ -12,5 +16,14 @@ func StringPartition[T any](values []discrete.Value, items []T) discrete.Solutio
 		groups := PartitionStrings(solution, values, items)
 		partition := sortedPartitionGroups(groups)
 		return strings.Join(partition, " ")
+	}
+}
+
+// StringSubset displays the solution as subset
+func StringSubset[T cmp.Ordered](items []T) discrete.SolutionStringFn {
+	return func(solution *discrete.Solution) string {
+		subset := list.MapList(AsSubset(solution), items)
+		slices.Sort(subset)
+		return str.WrapBraces(list.Map(subset, str.Any))
 	}
 }