+++ problem/problem_common.go
@@ -22,6 +22,25 @@ func newBinProblem(name string) (*discrete.Problem, *data.Bins) {
 	p.ObjectiveFn = fn.ScoreCountUniqueValues
 	p.SolutionCoreFn = fn.CoreSortedPartition(cfg.Bins, cfg.Weight)
 	p.SolutionStringFn = fn.StringPartition(cfg.Bins, cfg.Weight)
-
 	return p, cfg
+}
+
+// Common steps for creating Graph Cover problems
+func newGraphCoverProblem(name string, variablesFn data.GraphVariablesFn) (*discrete.Problem, *data.Graph) {
+	graph := data.NewUndirectedGraph(name)
+	if graph == nil {
+		return nil, nil
+	}
+
+	p := discrete.NewProblem(name)
+	p.Type = discrete.Subset
+
+	variables := variablesFn(graph)
+	p.Variables = discrete.Variables(variables)
+	p.AddVariableDomains(discrete.BooleanDomain())
+
+	p.Goal = discrete.Minimize
+	p.ObjectiveFn = fn.ScoreSubsetSize
+	p.SolutionStringFn = fn.StringSubset(variables)
+	return p, graph
 }