+++ problem/problem_registry.go
@@ -7,13 +7,19 @@ import (
 )
 
 const (
-	BinCover   = "bin_cover"
-	BinPacking = "bin_packing"
+	BinCover    = "bin_cover"
+	BinPacking  = "bin_packing"
+	CliqueCover = "clique_cover"
+	EdgeCover   = "edge_cover"
+	VertexCover = "vertex_cover"
 )
 
 var Creator = map[string]func(string, int) *discrete.Problem{
-	BinCover:   NewBinCover,
-	BinPacking: NewBinPacking,
+	BinCover:    NewBinCover,
+	BinPacking:  NewBinPacking,
+	CliqueCover: NewCliqueCover,
+	EdgeCover:   NewEdgeCover,
+	VertexCover: NewVertexCover,
 }
 
 // Create problem test case name