+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.3.18 - Graph Cover Problems
+  * **Commit**: 2026-01-31 18:21
+  * problem.NewVertexCover 
+  * problem.NewEdgeCover
+  * problem.NewCliqueCover
+  * fn.IsClique
 ## v0.3.17 - Bin Problems
   * **Commit**: 2026-01-31 17:45
   * problem.NewBinCover